Split out job.py, config.py. Remove _etconf.py dependency. Related encapsulation changes.
