+++ tests/conftest.py
@@ -10,7 +10,7 @@ pytest_plugins = [
     "fixtures._compute_resource",
     "fixtures._kachery_server",
     "fixtures._mongodb"
-]   
+]
 
 def pytest_addoption(parser):
     parser.addoption('--container', action='store_true', dest="container",
