+++ tests/test_remote.py
@@ -6,11 +6,14 @@ from .fixtures import MONGO_PORT, DATABASE_NAME, COMPUTE_RESOURCE_ID
 
 @pytest.mark.remote
 def test_remote_1(general, mongodb, kachery_server, compute_resource):
+    print("Entered test")
     db = hi.Database(mongo_url=f'mongodb://localhost:{MONGO_PORT}', database=DATABASE_NAME)
     jh = hi.RemoteJobHandler(database=db, compute_resource_id=COMPUTE_RESOURCE_ID)
     with hi.config(job_handler=jh, container=True):
         a = fun.ones.run(shape=(4, 3))
+        print("Point A")
         a = a.wait()
+        print("Point B")
         assert np.array_equal(a, np.ones((4, 3)))
         assert jh._internal_counts.num_jobs == 1, f'Unexpected number of jobs: {jh._internal_counts.num_jobs}'
